+++ src/components/GeneralComponents/SearchBox/index.js
@@ -2,7 +2,10 @@ import SearchIcon from "@mui/icons-material/Search";
 
 import { useState, useContext } from "react";
 
+import { useNavigate } from "react-router";
+
 import CurrentBook from "../../../context/CurrentBook";
+import BibleResults from "../../../context/BibleResults";
 
 import ShowVerse from "../../BibleComponents/ShowVerse";
 
@@ -11,10 +14,11 @@ import { Search, SearchIconWrapper, StyledInputBase } from "./styles";
 import { BASE_URL } from "../../../configs";
 
 function SearchBox() {
+  const navigate = useNavigate();
   const { currentBook } = useContext(CurrentBook);
+  const { setResults } = useContext(BibleResults);
 
   const [input, setInput] = useState();
-  const [results, setResults] = useState();
 
   function searchPassages(e) {
     e.preventDefault();
@@ -29,6 +33,7 @@ function SearchBox() {
         const data = xhr.response;
         console.log(JSON.parse(data));
         setResults(JSON.parse(data));
+        navigate("pesquisa/");
       } else {
         console.log(`Error: ${xhr.status}`);
       }
@@ -49,10 +54,6 @@ function SearchBox() {
           inputProps={{ "aria-label": "search" }}
         />
       </form>
-      {results &&
-        results.map((result) => {
-          return <ShowVerse verse={result} />;
-        })}
     </Search>
   );
 }
