+++ src/context/BibleResults.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const BibleResults = createContext({});
+
+export default BibleResults;
